+++ tests/main/test_main_csv.py
@@ -1,0 +1,61 @@
+from argparse import Namespace
+from pathlib import Path
+from tempfile import TemporaryDirectory
+
+import pytest
+from freezegun import freeze_time
+
+from datamodel_code_generator.__main__ import Exit, main
+from tests.main.test_main_general import DATA_PATH, EXPECTED_MAIN_PATH
+
+MonkeyPatch = pytest.MonkeyPatch
+
+CSV_DATA_PATH: Path = DATA_PATH / 'csv'
+
+
+@pytest.fixture(autouse=True)
+def reset_namespace(monkeypatch: MonkeyPatch):
+    namespace_ = Namespace(no_color=False)
+    monkeypatch.setattr('datamodel_code_generator.__main__.namespace', namespace_)
+    monkeypatch.setattr('datamodel_code_generator.arguments.namespace', namespace_)
+
+
+@freeze_time('2019-07-26')
+def test_csv_file():
+    with TemporaryDirectory() as output_dir:
+        output_file: Path = Path(output_dir) / 'output.py'
+        return_code: Exit = main(
+            [
+                '--input',
+                str(CSV_DATA_PATH / 'simple.csv'),
+                '--output',
+                str(output_file),
+                '--input-file-type',
+                'csv',
+            ]
+        )
+        assert return_code == Exit.OK
+        assert (
+            output_file.read_text()
+            == (EXPECTED_MAIN_PATH / 'csv_file_simple' / 'output.py').read_text()
+        )
+
+
+@freeze_time('2019-07-26')
+def test_csv_stdin(monkeypatch):
+    with TemporaryDirectory() as output_dir:
+        output_file: Path = Path(output_dir) / 'output.py'
+        monkeypatch.setattr('sys.stdin', (CSV_DATA_PATH / 'simple.csv').open())
+        return_code: Exit = main(
+            [
+                '--output',
+                str(output_file),
+                '--input-file-type',
+                'csv',
+            ]
+        )
+        assert return_code == Exit.OK
+        assert (
+            output_file.read_text()
+            == (EXPECTED_MAIN_PATH / 'csv_stdin_simple' / 'output.py').read_text()
+        )
